Authentication using Chrome localStorage

+++ Authorization Testing No - 2/my-auth-react-app-1/src/Components/Signup.jsx
@@ -9,10 +9,20 @@ export default function Singup()
 
     const addUser = () =>
     {
-        const userCount = localStorage.length
-        const nextIndex = userCount + 1
 
-        localStorage.setItem(`user${nextIndex}`, JSON.stringify({ username, password }))
+        const userExist = Object.values(localStorage).some((storedUser) => {
+            const user = JSON.parse(storedUser)
+            return user.username === username
+        })
+
+        if(!userExist) {
+            const userCount = localStorage.length
+            const nextIndex = userCount + 1
+            localStorage.setItem(`user${nextIndex}`, JSON.stringify({ username, password }))
+            alert(`user added : ${username}`)
+        } else {
+            alert(`this passowrd : ${password} this username :${username} has it so you cannot SignUP`)
+        }
     }
 
     return <div className=" flex gap-3 flex-col text-black">
@@ -22,18 +32,21 @@ export default function Singup()
         <input
         className="p-4"
         type="text"
-        placeholder="Username"
+        placeholder="Enter your Username"
         value={username}
         onChange={(e) => setUsername(e.target.value)}
          />
         <input
         className="p-4"
         type="text"
-        placeholder="Password"
+        placeholder="Enter your Password"
         value={password}
         onChange={(e) => setPassword(e.target.value)}
          />
 
          <button onClick={addUser} className=" p-4 bg-fuchsia-500 text-white font-semibold text-xl">Submit - signUp</button>
+
+        <a href="http://localhost:5173/Login" className="p-4 bg-lime-500 text-white font-semibold">To Login-Page</a>
+
     </div>
 }
